+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/descriptors/AttributeDescriptor.java
@@ -47,8 +47,6 @@ public class AttributeDescriptor implements Descriptor {
     @JsonProperty("@sql.foreign")
     private SqlForeign sqlForeign;
 
-
-
     @JsonProperty("@value")
     public String getValue() {
         return value;